Fix broken lambdas and get rid of multiline implementation
Multiline input implementation I offered actually broke the lambda
support because I was blindly replacing "\\" with " ". I should find
a better way to do this, but for now I will get rid of that part.
One can still copy-paste a formated input that contains "\n" and
that should still work.