Add initial support for floating point numbers
Treat all operations as floating point if at least one operand is.
Treat division as a floating point operation.
Change methods "as_*()" for integral types to return the actual value
and not a reference - the code looks simpler this way.

In the future this will make operations like "sqrt" possible.

Example:
  ~# lispy-rs
  lispy> + 1.0 2.0
  3
  lispy> + 1.0 2.1
  3.1
  lispy> + 1 2.1
  3.1
  lispy> / 1 2
  0.5